+++ src/rmcs_analyzer/export/__init__.py
@@ -1,6 +1,7 @@
 """Export services for RMCS Analyzer."""
 
 from .burnsim import export_burnsim_csv
+from .rasp import export_rasp_eng
 from .pdf_report import (
     PDFReportError,
     generate_campaign_pdf_report,
@@ -12,4 +13,5 @@ __all__ = [
     "generate_campaign_pdf_report",
     "generate_pdf_report",
     "export_burnsim_csv",
+    "export_rasp_eng",
 ]